requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
jinja2==3.1.4
pillow==10.4.0
python-multipart==0.0.9
transformers==4.44.2
torch>=2.1.0
accelerate==0.34.2
safetensors==0.4.4
diffusers==0.30.3
sentencepiece==0.2.0
